Modifique lo del frontEnd, ademas la ventana emergente esta centrada y deberia tener siempre el tamaño del tablero.

+++ src/game/Minimax.java
@@ -40,7 +40,7 @@ public class Minimax {
 		int depth=1;
 		Move move=null,auxMove;	
 		while(System.currentTimeMillis()<timeBound){
-			System.out.println(timeBound-System.currentTimeMillis()+" "+depth);
+	//		System.out.println(timeBound-System.currentTimeMillis()+" "+depth);
 			Node start=null;
 			if(tree){
 				try {
@@ -50,7 +50,7 @@ public class Minimax {
 				}
 				start=new Node();
 			}
-			auxMove=minimax(state,depth++,prune,null,timeBound);
+			auxMove=minimax(state,depth++,prune,start,timeBound);
 			if(tree){
 				start.setLabel("START "+move.getValue());
 				start.setColor("salmon");
@@ -63,13 +63,15 @@ public class Minimax {
 			if(auxMove!=null){
 				move=auxMove;
 				//Renombro
-				Node.rename("treeAux.dot","tree.dot");
+				if(tree) Node.rename("treeAux.dot","tree.dot");
 			}
 		}
-		//elimino el treeAux
-		File closeFile = new File("treeAux.dot");
-		closeFile.delete();
-		System.out.println(timeBound-System.currentTimeMillis()+" "+depth);
+		if(tree){
+			//elimino el treeAux
+			File closeFile = new File("treeAux.dot");
+			closeFile.delete();
+		}
+	//	System.out.println(timeBound-System.currentTimeMillis()+" "+depth);
 		return move;
 	}
 	
@@ -151,106 +153,6 @@ public class Minimax {
 		return answer;
 	}
 	
-	/*
-	public static Move minimaxByTime(Game state, Integer depth, Integer prune, Node me, Long time){
-		Move move=null;
-		int auxDepth=1;
-		Long auxTime=time;
-		Move auxMove=null;
-		while(auxTime>0){ 
-			auxMove=miniMaxRecursive(state,auxDepth,(move==null)?Integer.MAX_VALUE:move.getValue(),null, auxTime);
-			if(move==null || auxMove.getValue()>move.getValue())
-				move=auxMove;
-			auxTime=auxMove.getTime();
-			System.out.println(auxDepth+" time: "+auxTime);
-			auxDepth++;
-		}
-		return move;
-	}
-	
-	//si bien el tiempo lo pasan en segundos, apenas lo parseamos en el main lo pasamos a milisegundos 
-		private Move miniMaxRecursive(Game state, Integer depth, Integer prune, Node me, Long timeLeft){ 
-	 		if(depth==0 || state.getTurn()>2 ){
-	 			return new Move(state.value());
-	 		}
-	 		Move answer=new Move(Integer.MIN_VALUE);
-	 		Board board=state.getBoard();
-	 		List<Move> possibleMoves;
-	 		Integer actualPrune=null;
-	 		Node son=null,nodeAnswer=null;
-			Long timeUsed=null;Long initial=new Long(System.currentTimeMillis());
-			if(timeLeft!=null){
-				timeUsed=new Long(0);
-			}
-	 		if(prune!=null)
-	 			actualPrune=Integer.MAX_VALUE;
-	 		for(int i=0; i<board.getSize(); i++){
-	 			for(int j=0; j<board.getSize(); j++){
-	 				if(board.getPiece(i, j).getOwner()==state.getTurn()){
-	 					possibleMoves=getPossibleMovesFrom(board,i,j);
-	 					for (Move move : possibleMoves) {
-							if(timeLeft!=null && timeLeft <0){
-								answer.setTime(timeLeft);
-								return answer;
-							}
-							
-	 						Game stateAux= state.copy();
-	 						stateAux.move(move);
-	 						if(me!=null)
-	 							son=new Node();
-							if(timeLeft!=null){
-								if(timeLeft <0){
-									answer.setTime(timeLeft);
-									return answer;
-								}else{
-									timeUsed= System.currentTimeMillis()-initial;
-								}
-							}
-							//System.out.println(blancos(4-depth)+"Entre: "+move);
-							Move resp=miniMaxRecursive(stateAux,depth-1,actualPrune,son,timeLeft-timeUsed);
-							move.setValue(-resp.getValue());	
-							timeUsed= System.currentTimeMillis()-initial;
-							move.setTime(timeLeft-(System.currentTimeMillis()-initial));
-							//System.out.println(blancos(4-depth)+"Sali: "+move);
-	 						if(son!=null){
-	 							son.setMove(move);
-	 							if(depth%2==0)
-	 								son.setForm("ellipse");
-	 							else
-	 								son.setForm("box");
-	 							me.link(son);
-	 						}
-	 						
-	 						if (move.getValue()>answer.getValue()){							
-	 							if(me!=null){
-	 								nodeAnswer=son;
-	 							}
-	 							answer=move;
-	 							if(answer.getValue()==Integer.MAX_VALUE){
-	 								return answer;
-	 							}							
-	 						}
-	 						if(prune!=null){ //si en vez de un for each por los moves hago un for comun, lo que puedo hacer aca adentro 
-	 							if(move.getValue()>=prune){//es recorrer los nodos que me faltan antes de hacer el break y pintarlos
-	 								return answer;			//como nodos podados, para no gastar tanta memoria en todos los moves 
-	 							}else{						//haciendo el for each afuera
-	 								actualPrune=-answer.getValue();
-	 							}
-	 						}
-							if(timeUsed!=null && timeUsed>=timeLeft){
-								answer.setTime(timeLeft-(System.currentTimeMillis()-initial));
-								return answer;
-							}
-	 					}
-	 				}
-	 			}
-	 		}
-	 		if(nodeAnswer!=null) nodeAnswer.setColor("salmon");
-			answer.setTime(timeLeft-(System.currentTimeMillis()-initial));
-	 		return answer;
-	 	}
-	*/
-
 	//para debugear
 	private static String blancos(int a){
 		String b="";
